Made MojoRelation take different types for child and parent
There was no reason for child and parent in a MojoRelation to be of the
same type. The template now takes two types: one for the child and one
for the parent. To get the old behavior, specify the same type twice.
For use in set functions, both types still need to be the same.